Started refactoring tests, completed container tests
