+++ pkg/providers/cluster.go
@@ -3,9 +3,9 @@ package providers
 import (
 	"errors"
 
+	hclog "github.com/hashicorp/go-hclog"
 	"github.com/shipyard-run/shipyard/pkg/clients"
 	"github.com/shipyard-run/shipyard/pkg/config"
-	hclog "github.com/hashicorp/go-hclog"
 )
 
 var (
@@ -22,7 +22,7 @@ type Cluster struct {
 }
 
 // NewCluster creates a new
-func NewCluster(c *config.Cluster, cc clients.Docker, kc clients.Kubernetes,l hclog.Logger) *Cluster {
+func NewCluster(c *config.Cluster, cc clients.Docker, kc clients.Kubernetes, l hclog.Logger) *Cluster {
 	return &Cluster{c, cc, kc, l}
 }
 
@@ -52,14 +52,17 @@ func (c *Cluster) Destroy() error {
 
 // Lookup the a clusters current state
 func (c *Cluster) Lookup() (string, error) {
-	// lookup the server id
-	// base of cluster is a container
-	co := &config.Container{
-		Name:       c.config.Name,
-		NetworkRef: c.config.NetworkRef,
-	}
+	/*
+		// lookup the server id
+		// base of cluster is a container
+		co := &config.Container{
+			Name:       c.config.Name,
+			NetworkRef: c.config.NetworkRef,
+		}
 
-	p := NewContainer(co, c.client, c.log.With("parent_ref", c.config.Name))
+		p := NewContainer(co, c.client, c.log.With("parent_ref", c.config.Name))
 
-	return p.Lookup()
+		return p.Lookup()
+	*/
+	return "", nil
 }
