+++ pkg/providers/exec_remote.go
@@ -1,12 +1,9 @@
 package providers
 
 import (
-	"fmt"
-
 	hclog "github.com/hashicorp/go-hclog"
 	"github.com/shipyard-run/shipyard/pkg/clients"
 	"github.com/shipyard-run/shipyard/pkg/config"
-	"golang.org/x/xerrors"
 )
 
 // RemoteExec provider allows the execution of arbitrary commands on an existing target or
@@ -24,95 +21,102 @@ func NewRemoteExec(c *config.RemoteExec, ex clients.Docker, l hclog.Logger) *Rem
 
 // Create a new execution instance
 func (c *RemoteExec) Create() error {
-	c.log.Info("Remote executing command", "ref", c.config.Name, "command", c.config.Command, "args", c.config.Arguments, "image", c.config.Image)
+	/*
+		c.log.Info("Remote executing command", "ref", c.config.Name, "command", c.config.Command, "args", c.config.Arguments, "image", c.config.Image)
 
-	if c.config.Script != "" {
-		return fmt.Errorf("Remote execution of Scripts are not currently implemented: %s", c.config.Script)
-	}
-
-	// execution target id
-	targetID := ""
-	var targetContainer *Container // created if necessary
-
-	if c.config.TargetRef == nil {
-		// Not using existing target create new container
-		id, cp, err := c.createRemoteExecContainer()
-		if err != nil {
-			return err
+		if c.config.Script != "" {
+			return fmt.Errorf("Remote execution of Scripts are not currently implemented: %s", c.config.Script)
 		}
 
-		targetID = id
-		targetContainer = cp
-	} else {
-		// Fetch the id for the target
-		switch v := c.config.TargetRef.(type) {
-		case *config.Container:
-			cc := &config.Container{
-				Name:       v.Name,
-				Network:    v.Network,
-				NetworkRef: v.NetworkRef,
-			}
-			cp := NewContainer(cc, c.client, c.log)
-			id, err := cp.Lookup()
+		// execution target id
+		targetID := ""
+		var targetContainer *Container // created if necessary
 
+		if c.config.TargetRef == nil {
+			// Not using existing target create new container
+			id, cp, err := c.createRemoteExecContainer()
 			if err != nil {
-				return xerrors.Errorf("Unable to find remote exec target: %w", err)
-			}
-
-			if id == "" {
-				return xerrors.Errorf("Unable to find remote exec target")
+				return err
 			}
 
 			targetID = id
+			targetContainer = cp
+		} else {
+			// Fetch the id for the target
+			switch v := c.config.TargetRef.(type) {
+			case *config.Container:
+				cc := &config.Container{
+					Name:       v.Name,
+					Network:    v.Network,
+					NetworkRef: v.NetworkRef,
+				}
+				cp := NewContainer(cc, c.client, c.log)
+				id, err := cp.Lookup()
+
+				if err != nil {
+					return xerrors.Errorf("Unable to find remote exec target: %w", err)
+				}
+
+				if id == "" {
+					return xerrors.Errorf("Unable to find remote exec target")
+				}
+
+				targetID = id
+			}
 		}
-	}
 
-	// execute the script in the container
-	command := []string{}
-	command = append(command, c.config.Command)
-	command = append(command, c.config.Arguments...)
-	err := execCommand(c.client, targetID, command, c.log)
-	if err != nil {
-		return xerrors.Errorf("Unable to execute command in remote container: %w", err)
-	}
+		// execute the script in the container
+		command := []string{}
+		command = append(command, c.config.Command)
+		command = append(command, c.config.Arguments...)
+		err := execCommand(c.client, targetID, command, c.log)
+		if err != nil {
+			return xerrors.Errorf("Unable to execute command in remote container: %w", err)
+		}
 
-	// destroy the container if we created one
-	if c.config.Target == "" {
-		return targetContainer.Destroy()
-	}
+		// destroy the container if we created one
+		if c.config.Target == "" {
+			return targetContainer.Destroy()
+		}
+
+	*/
 
 	return nil
 }
 
 func (c *RemoteExec) createRemoteExecContainer() (string, *Container, error) {
-	// first create a new container
-	cc := &config.Container{
-		Name:        "remote_exec_temp",
-		Image:       *c.config.Image,
-		Command:     []string{"tail", "-f", "/dev/null"}, // ensure container does not immediately exit
-		Volumes:     c.config.Volumes,
-		NetworkRef:  c.config.WANRef, // seems like wan connections are not implemented //TODO fix
-		Environment: c.config.Environment,
-	}
+	/*
+		// first create a new container
+		cc := &config.Container{
+			Name:        "remote_exec_temp",
+			Image:       *c.config.Image,
+			Command:     []string{"tail", "-f", "/dev/null"}, // ensure container does not immediately exit
+			Volumes:     c.config.Volumes,
+			NetworkRef:  c.config.WANRef, // seems like wan connections are not implemented //TODO fix
+			Environment: c.config.Environment,
+		}
 
-	cp := NewContainer(cc, c.client, c.log)
+		cp := NewContainer(cc, c.client, c.log)
 
-	err := cp.Create()
-	if err != nil {
-		return "", nil, xerrors.Errorf("Unable to create container for remote exec: %w", err)
-	}
+		err := cp.Create()
+		if err != nil {
+			return "", nil, xerrors.Errorf("Unable to create container for remote exec: %w", err)
+		}
 
-	// get the id of the new container
-	id, err := cp.Lookup()
-	if err != nil {
-		return "", nil, xerrors.Errorf("Unable to find id for remote exec container: %w", err)
-	}
+		// get the id of the new container
+		id, err := cp.Lookup()
+		if err != nil {
+			return "", nil, xerrors.Errorf("Unable to find id for remote exec container: %w", err)
+		}
 
-	if id == "" {
-		return "", nil, xerrors.Errorf("Unable to find id for remote exec container")
-	}
+		if id == "" {
+			return "", nil, xerrors.Errorf("Unable to find id for remote exec container")
+		}
 
-	return id, cp, nil
+		return id, cp, nil
+	*/
+
+	return "", nil, nil
 }
 
 // Destroy statisfies the interface requirements but is not used
