+++ pkg/providers/utils.go
@@ -4,18 +4,7 @@ import (
 	"fmt"
 	"os"
 	"strings"
-
-	"github.com/shipyard-run/shipyard/pkg/config"
 )
-
-// FQDN generate the name of a docker container
-func FQDN(name string, net *config.Network) string {
-	if net == nil {
-		return fmt.Sprintf("%s.shipyard", name)
-	}
-
-	return fmt.Sprintf("%s.%s.shipyard", name, net.Name)
-}
 
 // CreateKubeConfigPath creates the file path for the KubeConfig file when
 // using a K3s or Kind cluster driver
